Added status codes attributes for HTTP methods

+++ ClinicWebApplication.Web/Controllers/AppoinmentsController.cs
@@ -1,19 +1,19 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using ClinicWebApplication.DataLayer.Models;
-using Microsoft.EntityFrameworkCore;
 using ClinicWebApplication.Interfaces;
 using ClinicWebApplication.Web.ViewModels;
 using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using ClinicWebApplication.BusinessLayer.Services.InputValidationService;
 using ClinicWebApplication.BusinessLayer.Specification.AppoinmentSpecification;
+using ClinicWebApplication.BusinessLayer.Services.AuthenticationService;
 using Microsoft.Extensions.Logging;
 using System.Security.Claims;
-using ClinicWebApplication.BusinessLayer.Services.AuthenticationService;
 
 namespace ClinicWebApplication.Web.Controllers
 {
@@ -39,6 +39,10 @@ namespace ClinicWebApplication.Web.Controllers
             var appoinments = await _appoinmentRepository.GetAll();
             return _mapper.Map<IEnumerable<Appoinment>, IEnumerable<AppoinmentViewModel>>(appoinments);
         }
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpGet("{id}")]
         [Authorize(Roles = "Doctor, Admin")]
         public async Task<ActionResult<AppoinmentViewModel>> Get(int id)
@@ -51,6 +55,9 @@ namespace ClinicWebApplication.Web.Controllers
             var appoinmentViewModel = _mapper.Map<AppoinmentViewModel>(appoinment);
             return new ObjectResult(appoinmentViewModel);
         }
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPost]
         [Authorize(Roles = "Patient")]
         public async Task<ActionResult<Appoinment>> Post(Appoinment appoinment)
@@ -65,6 +72,10 @@ namespace ClinicWebApplication.Web.Controllers
 
             return Ok(appoinment);
         }
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpPut]
         [Authorize(Roles = "Patient, Doctor")]
         public async Task<ActionResult<Appoinment>> Put(Appoinment appoinment)
@@ -80,12 +91,18 @@ namespace ClinicWebApplication.Web.Controllers
 
             return Ok(appoinment);
         }
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [HttpDelete("{id}")]
         [Authorize(Roles = "Patient")]
         public async Task<ActionResult<Appoinment>> Delete(int id)
         {
             Appoinment appoinment = await _appoinmentRepository.GetById(id);
             if (appoinment == null) return NotFound();
+            if (User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role).Value == Role.Patient &&
+                Int32.Parse(User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value) != appoinment.PatientId) return BadRequest();
             await _appoinmentRepository.Delete(appoinment);
 
             _logger.LogInformation($"Patient \"{this.User.Identity.Name}[{User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier).Value}]\" " +
